refactor: 내부에서 쓰이는 영화 컴포넌트는 internal 폴더로 분리
