+++ src/components/movie/internal/movieItem.ts
